added custom hook for filter and Pagination
