+++ src/components/PaginationProducts/Pagination.scss
@@ -27,12 +27,6 @@
     &--active {
       background-color: variable.$primary;
     }
-    &--prev {
-    }
-    &--next {
-    }
-    &--disabled {
-    }
   }
   &__link {
     width: 100%;
@@ -47,10 +41,6 @@
     &--active {
       background-color: variable.$primary;
       color: variable.$white;
-    }
-    &--prev {
-    }
-    &--next {
     }
     &--disabled {
       color: variable.$secondary;
