+++ src/components/ProductsPage/ProductsPage.scss
@@ -55,8 +55,6 @@
     font-weight: 700;
     line-height: 1.5;
   }
-  &__pagination {
-  }
 }
 
 .select .color-secondary {
