+++ src/utils/functions/handleBackToTop.ts
@@ -1,11 +1,3 @@
-// export const handleBackToTop = () => {
-//   window.scrollTo({ top: 0, behavior: 'smooth' });
-// };
-
 export const handleBackToTop = () => {
-  const scrollToTop = () => {
-    window.scrollTo({ top: 0, behavior: 'smooth' });
-  };
-
-  requestAnimationFrame(scrollToTop);
+  window.scrollTo({ top: 0, behavior: 'smooth' });
 };
